Updated HomeController
Edited the output of println for the loading of sector yaml
Added code to load all the yaml files and fill the database with the content of the yaml files
Edited the call of the index with the information to be given.
Added mappers for the variables in the class Strike which are of the type List. This done as try catch.