Updated admin form
Changed the select field for the strikes back to html code
Split the company ownership in two divs to improve visual appeal
Function to add the strike id's to the select field
Function to get the selected strike as a Json, and fill the fields according to the information in the Json. Inline function to get the values of the different arrays in the json.