Updated the input and admin form
Added two images to show when no example of the article can be given.
When adding an article, the article extension is checked in orde to use the appropriate way of saving the file.
Admin form updated to look nicer.
Added three buttons on the button of adminform (for trial)
Search function added when looking for strike in inputfield
Updated the way to get the article, and how to display it.
Cleaned up the input form, the place where the info is given, how the days for the given month are calculated, and how the months are loaded in to the select.
Updated the routes to get the article from the given strike