Adjusted the form input and admin
Rewrote the way the input fields and the admin fields are shown. Made use of the helper.
Created a list of the days, years and duration to pass to the input form view.
Commented companynames, occupations and identityDetails in homecontroller as they are being stored in a slightly different way.
Rewrote in Strike the way companynames, occupations and identitydetails is being stored