Changed handling of helpers
Changed the names of the helpers to resemble the names provided in the strike object. This for autofilling the form according to the strike object given.
Changed the constructors of a couple of models for the new way of saving the input from the helpers.
Furthermore did some cleanup of the code.