Updated admin side and homecontroller
Check if an article is uploaded, if not the variable stays empty. Due to viewing the article downloading unvalid file.
Adjusted the search in admin form. This due to the possibility on looking by checked or unchecked and searching specific strike.
Furthermore standardized the image shown at the article field.
Added a new image for noArticle shown
Edited name of view.png to viewArticle.png