Updated the application
HomeController:
Added try catch around the code to save the strike.
Checks if the strike is valid and returns error messages to display on the page.
admin.scala:
Adjusted the way the page is shown
Added a logout button to the page
index.scala:
Changed message to multiple messages for the errors
Added Strike variable
Added logout button
Changed the way info is displayed in the selectize elements, this for when a Strike is inserted incorrectly and redisplayed.
Edited the Add Strike button
update.scala:
Edited the buttons