Processed the feedback from AFRICAS-7 in Jira
Removed duplicate code from AdminController and HomeController and placed it in a new class called StrikeController.
Relocated the Yaml.load functions to StrikeController and added a call to StrikeController on the places where the Yaml.load functions used to be.
Load the Yaml files as soon as the application is started and loads the index page with a check to see if it was already loaded.
Updated the class Filters to add a CSRFFilter to the application to prevent Cross Site Request Forgery.
Added CSRF tokens to all the html pages.
Removed ExampleFilter, Module, ApplicationTimer, AtomicCounter and Counter.
Removed cache and javaWS from libraryDependencies.
Removed the route to HomeController.getStrikes.
Removed web plugins and Play enhancer from plugins.sbt.