Fixed a bug on loading the yaml files and added a page.
Fixed a bug when loading the yaml files, as the files aren't loaded when directly surfing to the admin page. Addded a check to fix that.
Added a page to display the error when no strike is selected in the admin page, instead of a simple string in an empty page.
Removed unused imports and cleaned the code a bit by removing print lines.