Edited the way strike ids are given to the admin page
Send a list of strike ids to the admin page instead of a list of strikes.
Split the de-selection of the checked/unchecked buttons from the search function.
Added event handler to handle the pressing of the refresh button to prevent reload of the page. And made the button of type button to prevent the form from being submitted.
Added a route for the request of the ids.
Added a method in Strike to load the strike ids
Removed the check for 'refresh' in AdminController and added a method to get all strike ids