Updated the user overview
Created an useroverviewcontroller and took the code from admincontroller. So to seperate the actions from one another.
Added code to load the user in the fields on the page.
Added code to update the user with the newly selected rights and to logout and to navigate back to the admin page.
Shortened the code to get all the users as an array
Added new fields to the overview and made sure the name corresponds to the user model.
Added javascript code to load the user in the fields.
Added route for the update method and realigned all the code to keep it nice and clean