Update africa strikes for user login
Added button to the admin page to navigate to the user overview page.
Added checks on which button is pressed in admin
Rearranged the location of the method saveYamlFileToDatabase to StrikeController
Added check for the authority of the user and to handle the login  on authority.
Autofill the default users and authorities when no present.
Added buttons to update and/or delete the user with the corresponding actions.
When the user has no authority set, it is set to "Unauthorized"
Added model Authority
Added variable authority to user with corresponding methods
Edited class Authorized to fit the new needs
Added new Authorization handlers to the Controllers
Added error message div to login page when unauthorized
Imroved the useroverview page
Added default authorities to yml file
Added default users to yml file
SQL evolution updated
Added route to get all users
Split javascript code from the user overview to seperate file
Added new style to the default stylesheet.