Fixed bug with internet explorer caching
Added code to the methods that send back Json files to prevent the browser from caching these files.
Added code to get the text for when the user is unauthorized from the config file instead of hardcoded.
Removed the admins from the config file as they are set via the database.
Made sure the user is being reset to prevent sudden modification/deletion of a user.
Make sure the useroverview uses the noStrikeSelected form for when no user is selected because of Internet Explorer.
Added route for the badRequest with noStrikeSelected form.