Update
Made sure the subcriber is able to see his or her own input via the data overview page. This is done by checking whether the user is a subscriber on multiple methods. Made use of the already existing pages.
Furthermore cleaned up some text in buttons and added some comments to the code.