Update
Added check to the code to catch adjustments to the options via the developer tools. This prevents users from selecting strikes that don't belong to them.
Added a label to show whether someone has selected a strike that doesn't belong to them.
Auto hide the label (div) so it doesn't take up space.