Security update
Added code to check if the strike has already been approved server side.
Furthermore a check to see if the author information is the same as the username of the current user.
Also a check to prevent user from navigating to the update page when the strike has been approved already.
Added variables to the update page to display or hide buttons and to make the author information field readonly.
Renamed the id of the undo button to adjust the visibillity