Update to add label to the strike provided
Added Label model which can be added to the Strike.
Changed the way the order of transactions in Ebean are done due to adding new Labels to the database if not existent.
Added code to handle the adding of new Labels in StrikeController.
Small textual edit in IdentityDetail.
Added the Label variables to the class Strike.
Adjusted the views for admin, index and update to fit the new information for the Label.
Update on the database in the sql file.
Added code in the javascript to handle the label input field.