Update for the strike labels
Rewrite of the way the labels are added to the Strike. Improvement for when there are no labels given, the Strike is not saved.
Furthermore adjusted the way the article file name is received. This to fix the issue of it being unable to view the files added. 
And altered the code according to the tips Intellij gave.